refactor(web): 공통 컴포넌트 shared 폴더로 이동 - GNB, Footer 등 공통 컴포넌트를 src/components/shared로 이동 - 관련 import 경로 일괄 수정
