+++ apps/web/src/components/shared/Footer/index.tsx
